requirements.txt
flask
flask-cors
psycopg2-binary
google-generativeai
gunicorn
werkzeug
